admxgen: separate file handling from command
We will then move command handling to another directory.
